rendered all the isearchResults from the API and displayed it to the user. Still need to style the rendered list
